+++ owrx/controllers/settings/decoding.py
@@ -128,22 +128,30 @@ class DecodingSettingsController(SettingsFormController):
                     append="s",
                 ),
                 NumberInput(
-                    "vdl2_ttl",
-                    "VDL2 reports expiration time",
-                    validator=RangeValidator(30, 100000),
-                    append="s",
-                ),
-                NumberInput(
                     "hfdl_ttl",
                     "HFDL reports expiration time",
                     validator=RangeValidator(30, 100000),
                     append="s",
                 ),
                 NumberInput(
+                    "vdl2_ttl",
+                    "VDL2 reports expiration time",
+                    validator=RangeValidator(30, 100000),
+                    append="s",
+                ),
+                CheckboxInput(
+                    "vdl2_ignore_acks",
+                    "Filter out VDL2 acknowledgement messages",
+                ),
+                NumberInput(
                     "acars_ttl",
                     "ACARS reports expiration time",
                     validator=RangeValidator(30, 100000),
                     append="s",
+                ),
+                CheckboxInput(
+                    "acars_ignore_acks",
+                    "Filter out ACARS acknowledgement messages",
                 ),
             ),
             Section(