+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+**1.2.117**
+- Added ARINC622 payload parsing inside ACARS.
+- Added CPDLC payload parsing inside ACARS.
+- Added ADS-C payload parsing inside ACARS.
+- Added ACARS payload parsing inside VDL2.
+- Added ACARS payload parsing inside HFDL.
+- Added option to drop ACARS acknowledgements.
+- Added option to drop VDL2 acknowledgements.
+
 **1.2.116**
 - Added LoRa Meshtastic parser [Stanislav Lechev].
 - Added LoRa APRS parser [Stefan DC4HF].